+++ raptorWeb/authprofiles/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from authprofiles.models import UserProfileInfo, DiscordUserInfo
+
+admin.site.register(UserProfileInfo)
+admin.site.register(DiscordUserInfo)
