+++ raptorWeb/raptorWeb/settings.py
@@ -8,6 +8,7 @@ BASE_DIR = Path(__file__).resolve().parent.parent
 TEMPLATE_DIR = join(BASE_DIR, "templates")
 RAPTOMC_TEMPLATE_DIR = join(TEMPLATE_DIR, "raptormc")
 STAFFAPPS_TEMPLATE_DIR = join(TEMPLATE_DIR, 'staffapps')
+AUTH_TEMPLATE_DIR = join(TEMPLATE_DIR, 'authprofiles')
 APPLICATIONS_DIR = join(RAPTOMC_TEMPLATE_DIR, "applications")
 PROFILES_DIR = join(RAPTOMC_TEMPLATE_DIR, 'profiles')
 STATIC_DIR = join(BASE_DIR, "static")
@@ -56,6 +57,7 @@ INSTALLED_APPS = [
     'ckeditor',
     'raptormc',
     'staffapps',
+    'authprofiles'
 ]
 
 MIDDLEWARE = [
@@ -145,7 +147,7 @@ PASSWORD_HASHERS = [
 
 # Django Authentication
 AUTHENTICATION_BACKENDS = [
-    'raptormc.auth.DiscordAuthBackend',
+    'authprofiles.auth.DiscordAuthBackend',
     'django.contrib.auth.backends.ModelBackend'
 ]
 
@@ -222,6 +224,11 @@ LOGGING = {
             'propagate': False,
         },
         'staffapps.views': {
+            'handlers': ['console', 'log_file'],
+            'level': 'DEBUG',
+            'propagate': False,
+        },
+        'authprofiles.auth': {
             'handlers': ['console', 'log_file'],
             'level': 'DEBUG',
             'propagate': False,
