+++ raptorWeb/raptorWeb/urls.py
@@ -22,12 +22,14 @@ from django.conf.urls.static import static
 from raptorWeb.settings import RAPTOMC_TEMPLATE_DIR, DEBUG, MEDIA_URL, MEDIA_ROOT
 from raptormc import urls as SR_urls
 from staffapps import urls as app_urls
+from authprofiles import urls as auth_urls
 
 urlpatterns = [
 
     path('admin/', admin.site.urls, name="admin"),
     path('robots.txt', TemplateView.as_view(template_name=join(RAPTOMC_TEMPLATE_DIR, 'robots.txt'), content_type="text/plain")),
     path('staffapps/', include(app_urls), name="staffapps"),
+    path('', include(auth_urls), name="authprofiles"),
     path('', include(SR_urls), name="shadowraptormc"),
 
 ]
