+++ raptorWeb/raptormc/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.forms import TextInput, Textarea
 from django.db import models
-from raptormc.models import Server, InformativeText, PlayerCount, PlayerName, UserProfileInfo, DiscordUserInfo
+from raptormc.models import Server, InformativeText, PlayerCount, PlayerName
 
 class ServerAdmin(admin.ModelAdmin):
     formfield_overrides = {
@@ -19,5 +19,3 @@ admin.site.register(Server, ServerAdmin)
 admin.site.register(PlayerCount)
 admin.site.register(PlayerName)
 admin.site.register(InformativeText, InformativeTextAdmin)
-admin.site.register(UserProfileInfo)
-admin.site.register(DiscordUserInfo)
