+++ raptorWeb/templates/authprofiles/all_profiles.html
@@ -22,7 +22,7 @@
     <div class="row">
         {% for loaded_user in users %}
             <div class="col-md-4 col-12">
-                <a href="{% url 'raptormc:user_profile' loaded_user.username %}">
+                <a href="{% url 'authprofiles:user_profile' loaded_user.username %}">
                     <div id="profile_box" class="card bg-light m-3 p-3 opacity-75">
                         <div id="profile_picture_box" class="mx-auto d-block">
                             <img class="w-100" src="{{loaded_user.profile_picture}}" alt="">
