+++ raptorWeb/templates/authprofiles/login.html
@@ -4,14 +4,14 @@
 <link rel="stylesheet" href="{% static 'css/raptormc.css' %}">
 
 <div class="card bg-dark text-white p-1 mt-0 mb-0 opacity-75 container">
-  <form action="{% url 'raptormc:login' %}" method="POST">
+  <form action="{% url 'authprofiles:login' %}" method="POST">
 
     {% csrf_token %}
     {% bootstrap_form login_form %}
     <div class="loginSubmit opacity-100">
       {% bootstrap_button button_type="submit" content="Login" %}
     </div>
-    <a role="button" type="button" class="btn btn-discord mt-2 opacity-100" href="{% url 'raptormc:login_oauth' %}">Login with Discord</a>
+    <a role="button" type="button" class="btn btn-discord mt-2 opacity-100" href="{% url 'authprofiles:login_oauth' %}">Login with Discord</a>
 
   </form>
 </div>
