+++ raptorWeb/templates/authprofiles/profile.html
@@ -23,7 +23,7 @@
         {% if user.username == displayed_profile.base.username %}
             <div class="card bg-dark m-3 p-3">
                 <div class="row">
-                    <a role="button" href="{% url 'raptormc:user_profile_edit' user.username %}" class="btn btn-primary">
+                    <a role="button" href="{% url 'authprofiles:user_profile_edit' user.username %}" class="btn btn-primary">
                         Edit Profile Details
                     </a>
                 </div>
