+++ raptorWeb/templates/authprofiles/registration.html
@@ -26,7 +26,7 @@
     <section>
     
       <div class="card bg-light p-3 mt-4 opacity-75 container">
-        <a role="button" href="{% url 'raptormc:login_oauth' %}" class="btn btn-discord">
+        <a role="button" href="{% url 'authprofiles:login_oauth' %}" class="btn btn-discord">
           Register with Discord
         </a>
       </div>
