--- raptorWeb/templates/raptormc/applications/appsuccess.html
@@ -1,28 +1,0 @@
-{% extends 'raptormc/base.html' %}
-
-{% block page_title %}
-<title>
-  
-  ShadowRaptor | Application Success
-  
-</title>
-{% endblock %}
-{% block metadata %}
-<meta property="og:title" content="ShadowRaptor | Application Success" />
-{% endblock %}
-
-{% block main_block %}
-<section id="welcome" class="container">
-
-  <header class="card bg-dark opacity-75 text-white mt-4">
-    <h3 class="display-4 card-header fw-bold text-center">Application Successfully Submitted!</h3>
-    <p class="p-4 container">
-
-      <p class="pageHeader">Thank you for applying!</p>
-      <p class="pageContent">We appreciate your considering a staff position on our network! If we do approve your application, you will be notified via Discord. Be sure to watch for messages from unknown users!</p>
-
-    </p>
-  </header>
-
-</section>
-{% endblock %}
